[Test] Add Bal_library_url expansion variables
Issue #86

Similar to the current $bal_library_dir behaviour, add a
bal_library_dir_url which expands to a file url of the directory
the library is in.
Due to the desire to allow relative paths, we now invoke normalization
behaviour on substitutions depending on if the entire string is a file
url, ie, does it start with "file:".
These substitutions are compatible with openassetio url
conversion methods.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>